Поправлено получение меток на карте

+++ src/main/java/ru/leyman/loco/locoback/domain/dto/PostDto.java
@@ -3,7 +3,6 @@ package ru.leyman.loco.locoback.domain.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import ru.leyman.loco.locoback.domain.enums.PostCategory;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -11,5 +10,5 @@ import java.util.List;
 public record PostDto(Long id, Author author, LocalDateTime created,
                       String text, PostCategory category,
                       List<Long> contents,
-                      BigDecimal latitude, BigDecimal longitude) {
+                      Float latitude, Float longitude) {
 }
