+++ src/main/java/ru/leyman/loco/locoback/domain/dto/PostMark.java
@@ -1,9 +1,7 @@
 package ru.leyman.loco.locoback.domain.dto;
 
-import java.math.BigDecimal;
-
 /**
  * DTO для карты
  */
-public record PostMark(Long id, BigDecimal latitude, BigDecimal longitude) {
+public record PostMark(Long id, Float latitude, Float longitude) {
 }
